Fix multi-item handling.
I added code to add the "Send to SABnzbd" button at the top for nzbs.org
(I left the button-tweaking code unchanged, in case some site needs it).
Category parsing is busted, so I just put it in a try...catch.  I made
the tweaking of the button's value work with whatever "Send ..." string
is there (since the added button is "Send to SABnzbd", not "Send to
SAB").  Each sent item is unchecked, to match the way the other download
and add-to-cart buttons work.  Finally, I added "&del=1" to the fetch
URL so that any grabbed item is removed from the cart (which should
really be turned into an option, but I didn't try to do that yet).